sdlf-ddk-lightweight/requirements.txt
aws_ddk_core==0.4.0
boto3==1.21.4
aws-cdk-lib==2.20.0